Major refactor/oganization of codebase structure for web-page deployment
