--- postman_files/CS467 travel-companion.postman_environment.json
@@ -1,15 +1,0 @@
-{
-	"id": "a13b1a78-fb36-4588-acab-df5a4422f34c",
-	"name": "CS467 travel-companion",
-	"values": [
-		{
-			"key": "app_url",
-			"value": "http://127.0.0.1:8080",
-			"type": "default",
-			"enabled": true
-		}
-	],
-	"_postman_variable_scope": "environment",
-	"_postman_exported_at": "2024-07-20T17:11:34.364Z",
-	"_postman_exported_using": "Postman/11.1.25"
-}
